refactor: rename Campaign component
