+++ src/components/Campaigns/StepsImage/index.tsx
